+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+ - New Job Engine Lite Docker image build, optimized for running jobs from within an infrastructure, with multi-arch support
+ 
 ## [2.10.0] - 2020-12-10
 
 ### Added